Relocate individual modules' V3 API
