+++ cardinal-components-entity/src/main/java/nerdhub/cardinal/components/api/event/EntityComponentCallback.java
@@ -22,7 +22,7 @@
  */
 package nerdhub.cardinal.components.api.event;
 
-import dev.onyxstudios.cca.api.v3.component.entity.EntityComponentFactory;
+import dev.onyxstudios.cca.api.v3.entity.EntityComponentFactory;
 import dev.onyxstudios.cca.internal.entity.CardinalEntityInternals;
 import nerdhub.cardinal.components.api.ComponentType;
 import nerdhub.cardinal.components.api.component.Component;
