+++ cardinal-components-item/src/main/java/dev/onyxstudios/cca/api/v3/item/ItemComponentFactory.java
@@ -20,28 +20,44 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package dev.onyxstudios.cca.api.v3.component.level;
+package dev.onyxstudios.cca.api.v3.item;
 
-import net.minecraft.world.WorldProperties;
+import nerdhub.cardinal.components.api.component.Component;
+import nerdhub.cardinal.components.api.component.extension.CopyableComponent;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
 import org.jetbrains.annotations.ApiStatus;
 
+import javax.annotation.Nonnull;
+
+
 /**
- * Entrypoint getting invoked to register <em>static</em> item component factories.
+ * A component factory for {@linkplain ItemStack item stacks}.
  *
- * <p>The entrypoint is exposed as {@code cardinal-components-level} in the mod json and runs for any environment.
- * It usually executes right before the first {@linkplain WorldProperties save properties object} gets loaded.
+ * <p>When invoked, the factory must return a {@link Component} of the right type.
  *
  * @since 2.4.0
  */
 @ApiStatus.Experimental
-public interface LevelComponentInitializer {
+@FunctionalInterface
+public interface ItemComponentFactory<C extends CopyableComponent<?>> extends ItemComponentFactoryV2<C> {
+    @Nonnull
+    @Override
+    default C createForStack(Item item, ItemStack stack) {
+        return this.createForStack(stack);
+    }
+
     /**
-     * Called to register component factories for statically declared component types.
+     * Initialize components for the given stack.
      *
-     * <p><strong>The passed registry must not be held onto!</strong> Static component factories
-     * must not be registered outside of this method.
+     * <p>The component returned by this method will be available
+     * on the chunk as soon as all component factories have been invoked.
      *
-     * @param registry a {@link LevelComponentFactoryRegistry} for <em>statically declared</em> components
+     * @param stack      the {@code ItemStack} being constructed
+     * @implNote Because this method is called for each stack creation, implementations
+     * should avoid side effects and keep costly computations at a minimum. Lazy initialization
+     * should be considered for components that are costly to initialize.
      */
-    void registerLevelComponentFactories(LevelComponentFactoryRegistry registry);
+    @Nonnull
+    C createForStack(ItemStack stack);
 }
