+++ cardinal-components-level/src/main/java/dev/onyxstudios/cca/internal/level/StaticLevelComponentPlugin.java
@@ -23,9 +23,9 @@
 package dev.onyxstudios.cca.internal.level;
 
 import dev.onyxstudios.cca.api.v3.component.ComponentKey;
-import dev.onyxstudios.cca.api.v3.component.level.LevelComponentFactory;
-import dev.onyxstudios.cca.api.v3.component.level.LevelComponentFactoryRegistry;
-import dev.onyxstudios.cca.api.v3.component.level.LevelComponentInitializer;
+import dev.onyxstudios.cca.api.v3.level.LevelComponentFactory;
+import dev.onyxstudios.cca.api.v3.level.LevelComponentFactoryRegistry;
+import dev.onyxstudios.cca.api.v3.level.LevelComponentInitializer;
 import dev.onyxstudios.cca.internal.base.DynamicContainerFactory;
 import dev.onyxstudios.cca.internal.base.asm.StaticComponentPluginBase;
 import nerdhub.cardinal.components.api.component.Component;
