+++ cardinal-components-level/src/main/java/nerdhub/cardinal/components/api/event/LevelComponentCallback.java
@@ -22,7 +22,7 @@
  */
 package nerdhub.cardinal.components.api.event;
 
-import dev.onyxstudios.cca.api.v3.component.level.LevelComponentFactory;
+import dev.onyxstudios.cca.api.v3.level.LevelComponentFactory;
 import nerdhub.cardinal.components.api.ComponentType;
 import nerdhub.cardinal.components.api.component.Component;
 import nerdhub.cardinal.components.api.component.ComponentContainer;
