+++ cardinal-components-util/src/main/java/dev/onyxstudios/cca/api/v3/util/GenericComponentFactoryRegistry.java
@@ -20,36 +20,17 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package dev.onyxstudios.cca.api.v3.component.chunk;
+package dev.onyxstudios.cca.api.v3.util;
 
-import nerdhub.cardinal.components.api.component.Component;
-import nerdhub.cardinal.components.api.component.extension.CopyableComponent;
-import net.minecraft.world.chunk.Chunk;
+import com.google.common.reflect.TypeToken;
+import dev.onyxstudios.cca.api.v3.component.ComponentKey;
+import net.minecraft.util.Identifier;
 import org.jetbrains.annotations.ApiStatus;
 
-import javax.annotation.Nonnull;
-
 /**
- * A component factory for {@linkplain Chunk chunks}.
- *
- * <p>When invoked, the factory must return a {@link Component} of the right type.
- *
  * @since 2.4.0
  */
 @ApiStatus.Experimental
-@FunctionalInterface
-public interface ChunkComponentFactory<C extends CopyableComponent<?>> {
-    /**
-     * Initialize components for the given chunk.
-     *
-     * <p>The component returned by this method will be available
-     * on the chunk as soon as all component factories have been invoked.
-     *
-     * @param chunk      the chunk being constructed
-     * @implNote Because this method is called for each chunk creation, implementations
-     * should avoid side effects and keep costly computations at a minimum. Lazy initialization
-     * should be considered for components that are costly to initialize.
-     */
-    @Nonnull
-    C createForChunk(Chunk chunk);
+public interface GenericComponentFactoryRegistry {
+    <F> void register(ComponentKey<?> type, Identifier providerId, TypeToken<F> factoryType, F factory);
 }
