+++ cardinal-components-util/src/main/java/dev/onyxstudios/cca/internal/util/ComponentContainerMetafactoryImpl.java
@@ -26,7 +26,7 @@ import com.google.common.reflect.Invokable;
 import com.google.common.reflect.Parameter;
 import com.google.common.reflect.TypeToken;
 import dev.onyxstudios.cca.api.v3.component.ComponentContainer;
-import dev.onyxstudios.cca.api.v3.component.util.ContainerGenerationException;
+import dev.onyxstudios.cca.api.v3.util.ContainerGenerationException;
 import dev.onyxstudios.cca.internal.base.ComponentsInternals;
 import dev.onyxstudios.cca.internal.base.asm.CcaAsmHelper;
 import dev.onyxstudios.cca.internal.base.asm.StaticComponentLoadingException;
