+++ cardinal-components-util/src/main/java/dev/onyxstudios/cca/internal/util/StaticGenericComponentPlugin.java
@@ -25,8 +25,8 @@ package dev.onyxstudios.cca.internal.util;
 import com.google.common.reflect.TypeToken;
 import dev.onyxstudios.cca.api.v3.component.ComponentContainer;
 import dev.onyxstudios.cca.api.v3.component.ComponentKey;
-import dev.onyxstudios.cca.api.v3.component.util.GenericComponentFactoryRegistry;
-import dev.onyxstudios.cca.api.v3.component.util.GenericComponentInitializer;
+import dev.onyxstudios.cca.api.v3.util.GenericComponentFactoryRegistry;
+import dev.onyxstudios.cca.api.v3.util.GenericComponentInitializer;
 import dev.onyxstudios.cca.internal.base.LazyDispatcher;
 import dev.onyxstudios.cca.internal.base.asm.CcaAsmHelper;
 import dev.onyxstudios.cca.internal.base.asm.StaticComponentLoadingException;
