+++ cardinal-components-world/src/main/java/dev/onyxstudios/cca/api/v3/world/WorldComponentFactoryRegistry.java
@@ -20,39 +20,19 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package dev.onyxstudios.cca.api.v3.component.item;
+package dev.onyxstudios.cca.api.v3.world;
 
+import dev.onyxstudios.cca.api.v3.component.ComponentKey;
 import nerdhub.cardinal.components.api.component.Component;
-import nerdhub.cardinal.components.api.component.extension.CopyableComponent;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import org.jetbrains.annotations.ApiStatus;
-
-import javax.annotation.Nonnull;
-
 
 /**
- * A component factory for {@linkplain ItemStack item stacks}.
- *
- * <p>When invoked, the factory must return a {@link Component} of the right type.
- *
  * @since 2.4.0
  */
-@ApiStatus.Experimental
-@FunctionalInterface
-public interface ItemComponentFactoryV2<C extends CopyableComponent<?>> {
+public interface WorldComponentFactoryRegistry {
     /**
-     * Initialize components for the given stack.
+     * Registers a {@link WorldComponentFactory}.
      *
-     * <p>The component returned by this method will be available
-     * on the chunk as soon as all component factories have been invoked.
-     *
-     * @param item  the actual item held by the stack, may differ from the result of {@link ItemStack#getItem()}
-     * @param stack the {@code ItemStack} being constructed
-     * @implNote Because this method is called for each stack creation, implementations
-     * should avoid side effects and keep costly computations at a minimum. Lazy initialization
-     * should be considered for components that are costly to initialize.
+     * @param factory the factory to use to create components of the given type
      */
-    @Nonnull
-    C createForStack(Item item, ItemStack stack);
+    <C extends Component> void register(ComponentKey<C> type, WorldComponentFactory<? extends C> factory);
 }
