+++ cardinal-components-world/src/main/java/dev/onyxstudios/cca/api/v3/world/WorldComponentInitializer.java
@@ -20,27 +20,20 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package dev.onyxstudios.cca.api.v3.component.util;
+package dev.onyxstudios.cca.api.v3.world;
 
+import net.minecraft.world.World;
 import org.jetbrains.annotations.ApiStatus;
 
 /**
- * Entrypoint getting invoked to register <em>static</em> generic (typically for third party providers)
- * component factories.
+ * Entrypoint getting invoked to register <em>static</em> world component factories.
  *
- * <p>The entrypoint is exposed as {@code "cardinal-components-util"} in the mod json and runs for any environment.
+ * <p>The entrypoint is exposed as {@code cardinal-components-world} in the mod json and runs for any environment.
+ * It usually executes right before the first {@link World} instance is created.
  *
  * @since 2.4.0
  */
 @ApiStatus.Experimental
-public interface GenericComponentInitializer {
-    /**
-     * Called to register component factories for statically declared component types.
-     *
-     * <p><strong>The passed registry must not be held onto!</strong> Static component factories
-     * must not be registered outside of this method.
-     *
-     * @param registry a {@link GenericComponentFactoryRegistry} for <em>statically declared</em> components
-     */
-    void registerGenericComponentFactories(GenericComponentFactoryRegistry registry);
+public interface WorldComponentInitializer {
+    void registerWorldComponentFactories(WorldComponentFactoryRegistry registry);
 }
