+++ cardinal-components-world/src/main/java/dev/onyxstudios/cca/internal/world/StaticWorldComponentPlugin.java
@@ -23,9 +23,9 @@
 package dev.onyxstudios.cca.internal.world;
 
 import dev.onyxstudios.cca.api.v3.component.ComponentKey;
-import dev.onyxstudios.cca.api.v3.component.world.WorldComponentFactory;
-import dev.onyxstudios.cca.api.v3.component.world.WorldComponentFactoryRegistry;
-import dev.onyxstudios.cca.api.v3.component.world.WorldComponentInitializer;
+import dev.onyxstudios.cca.api.v3.world.WorldComponentFactory;
+import dev.onyxstudios.cca.api.v3.world.WorldComponentFactoryRegistry;
+import dev.onyxstudios.cca.api.v3.world.WorldComponentInitializer;
 import dev.onyxstudios.cca.internal.base.asm.StaticComponentPluginBase;
 import nerdhub.cardinal.components.api.component.Component;
 import net.fabricmc.loader.api.FabricLoader;
